+++ foxes/models/wake_frames/yawed_wakes.py
@@ -36,7 +36,7 @@ class YawedWakes(WakeFrame):
         self,
         k=None,
         ct_max=0.9999,
-        alpha=0.58,
+        alpha=0.58, 
         beta=0.07,
         base_frame=RotorWD(),
         k_var=FV.K,