+++ foxes/models/wake_models/wind/porte_agel.py
@@ -1,16 +1,18 @@
 import numpy as np
 
 from foxes.models.wake_models.dist_sliced import DistSlicedWakeModel
+from foxes.core.model import Model
 import foxes.variables as FV
 import foxes.constants as FC
 
 
-class PorteAgelModel:
+class PorteAgelModel(Model):
     """
     Common calculations for the wake model and the wake
     frame, such that code repetitions can be avoided.
 
-    Based on Bastankhah & Porte-Agel, 2016, https://doi.org/10.1017/jfm.2016.595
+    Based on Bastankhah & Porte-Agel, 2016, 
+    https://doi.org/10.1017/jfm.2016.595
 
     Attributes
     ----------
@@ -21,6 +23,8 @@ class PorteAgelModel:
         model parameter used to determine onset of far wake region
     beta: float
         model parameter used to determine onset of far wake region
+    
+    :group: models.wake_models.wind
 
     """
 
@@ -56,9 +60,10 @@ class PorteAgelModel:
             model parameter used to determine onset of far wake region
 
         """
+        super().__init__()
         self.ct_max = ct_max
-        self.alpha = alpha
-        self.beta = beta
+        setattr(self, FV.PA_ALPHA, alpha)
+        setattr(self, FV.PA_BETA, beta)
 
     @property
     def pars(self):
@@ -71,7 +76,9 @@ class PorteAgelModel:
             Dictionary of the model parameters
 
         """
-        return dict(alpha=self.alpha, beta=self.beta, ct_max=self.ct_max)
+        alpha = getattr(self, FV.PA_ALPHA)
+        beta = getattr(self, FV.PA_BETA)
+        return dict(alpha=alpha, beta=beta, ct_max=self.ct_max)
 
     def calc_data(
         self,
@@ -107,6 +114,7 @@ class PorteAgelModel:
         n_points = x.shape[1]
         st_sel = (np.arange(n_states), states_source_turbine)
 
+
         # store parameters:
         out = {self.PARS: self.pars}
         out[self.CHECK] = (
@@ -135,6 +143,14 @@ class PorteAgelModel:
             ti = np.zeros((n_states, n_points), dtype=FC.DTYPE)
             ti[:] = fdata[FV.TI][st_sel][:, None]
 
+            # get alpha:
+            alpha = np.zeros((n_states, n_points), dtype=FC.DTYPE)
+            alpha[:] = Model.get_data(self, FV.PA_ALPHA, fdata, data_prio=True, upcast="farm")[st_sel][:, None]
+
+            # get beta:
+            beta = np.zeros((n_states, n_points), dtype=FC.DTYPE)
+            beta[:] = Model.get_data(self, FV.PA_BETA, fdata, data_prio=True, upcast="farm")[st_sel][:, None]
+
             # apply filter:
             x = x[sp_sel]
             D = D[sp_sel]
@@ -143,14 +159,14 @@ class PorteAgelModel:
             ti = ti[sp_sel]
             k = k[sp_sel]
             gamma = gamma[sp_sel]
-
+            alpha = alpha[sp_sel]
+            beta = beta[sp_sel]
+        
             # calc theta_c0, Eq. (6.12):
             cosg = np.cos(gamma)
             theta = 0.3 * gamma / cosg * (1 - np.sqrt(1 - ct * cosg))
 
             # calculate x0, Eq. (7.3):
-            alpha = self.alpha
-            beta = self.beta
             sqomct = np.sqrt(1 - ct)
             x0 = (
                 D
@@ -294,7 +310,8 @@ class PorteAgelWake(DistSlicedWakeModel):
     """
     The Bastankhah PorteAgel wake model
 
-    Based on Bastankhah & Porte-Agel, 2016, https://doi.org/10.1017/jfm.2016.595
+    Based on Bastankhah & Porte-Agel, 2016, 
+    https://doi.org/10.1017/jfm.2016.595
 
     Attributes
     ----------
@@ -313,7 +330,15 @@ class PorteAgelWake(DistSlicedWakeModel):
 
     """
 
-    def __init__(self, superposition, k=None, ct_max=0.9999, alpha=0.58, beta=0.07, k_var=FV.K):
+    def __init__(
+            self, 
+            superposition, 
+            k=None, 
+            ct_max=0.9999, 
+            alpha=0.58, 
+            beta=0.07, 
+            k_var=FV.K,
+        ):
         """
         Constructor.